add discharge method readme

+++ src/bms_can_utils.h
@@ -82,10 +82,10 @@ namespace can
             {
                 std::array<int16_t, 4>
                     data = {
-                        temp_data.battery_average / 1000,
-                        temp_data.temp_range.first / 1000,
-                        temp_data.pcb_average / 1000,
-                        temp_data.temp_range.second / 1000};
+                        static_cast<int16_t>(temp_data.battery_average / 1000),
+                        static_cast<int16_t>(temp_data.temp_range.first / 1000),
+                        static_cast<int16_t>(temp_data.pcb_average / 1000),
+                        static_cast<int16_t>(temp_data.temp_range.second / 1000)};
                 transmit(protocol::create_packet_id(protocol::CAN_PACKET_ID::CAN_PACKET_BMS_STATUS_TEMPERATURES, board::CAN_ID), data);
             }
             if (request)
